requirements.txt
streamlit==1.41.1
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1
fredapi==0.5.2
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
statsmodels==0.14.4
scipy==1.13.1
openpyxl==3.1.5
xlrd==2.0.1
feedparser==6.0.11
pyyaml==6.0.2
pyarrow==18.1.0
